aggiunte condizioni delete custom

+++ src/main/java/it/prova/gestionetratte/service/AirbusServiceImpl.java
@@ -6,9 +6,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import it.prova.gestionetratte.dto.AirbusDTO;
 import it.prova.gestionetratte.model.Airbus;
 import it.prova.gestionetratte.repository.airbus.AirbusRepository;
 import it.prova.gestionetratte.web.api.exception.AirbusNotFoundException;
+import it.prova.gestionetratte.web.api.exception.AirbusWithTratteException;
 
 @Service
 public class AirbusServiceImpl implements AirbusService {
@@ -46,8 +48,17 @@ public class AirbusServiceImpl implements AirbusService {
 	public void rimuovi(Long idToRemove) {
 		repository.findById(idToRemove)
 				.orElseThrow(() -> new AirbusNotFoundException("Airbus not found con id: " + idToRemove));
-
-		repository.deleteById(idToRemove);
+		//condizione airbus con tratte
+		AirbusDTO airbusToRemove =AirbusDTO.buildAirbusDTOFromModel(repository.findById(idToRemove).orElse(null),true);
+		try{
+			if(airbusToRemove.getTratte().size()>0)
+		         throw new AirbusWithTratteException("Airbus with tratte!");
+			
+			repository.deleteById(idToRemove);
+		}catch(AirbusWithTratteException e){
+			e.printStackTrace();
+			
+		}
 	}
 
 	public List<Airbus> findByExample(Airbus example) {
